refactor(ui): return the status colors applyTheme already resolved
statusColor re-parsed the theme's hex on every painted glyph while the
resolved colours sat unread, two of them with no reader at all. It returns
those now; statusHex stays for the one caller that mixes a hex string.

The selected-row test moves to styles_test.go, beside the function it
covers.
