+++ internal/ui/surface_test.go
@@ -19,12 +19,3 @@ func TestPaintReappliesFillAfterLipglossReset(t *testing.T) {
 		t.Fatalf("fill dropped after the inner reset: %q", got)
 	}
 }
-
-func TestRenderSelectedRowReappliesTintAfterLipglossReset(t *testing.T) {
-	inner := lipgloss.NewStyle().Bold(true).Render("name") + " rest"
-	got := renderSelectedRow(inner)
-	reapply := ansi.ResetStyle + bgSeq(current.Surface) + fgSeq(current.Bright)
-	if !strings.Contains(got, reapply+" rest") {
-		t.Fatalf("tint dropped after the inner reset: %q", got)
-	}
-}
